+++ packages/flame/CHANGELOG.md
@@ -7,6 +7,7 @@
  - Export new effects system
  - Introduce `updateTree` to follow the `renderTree` convention
  - Fix `Parallax.load` with different loading times
+ - `isHud` renamed to `respectCamera`
 
 ## [1.0.0-releasecandidate.18]
  - Forcing portrait and landscape mode is now supported on web